% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dscoemarkdown-package.R
\docType{package}
\name{dscoemarkdown}
\alias{dscoemarkdown}
\alias{dscoemarkdown-package}
\title{DSCOE Markdown}
\description{
This is the DSCOE format for knitr.
}
\details{
\itemize{

\item \code{dscoe} (\code{output: dscoemarkdown::dscoe}) uses the DSCOE-Modified Bulma CSS
framework and has optional \code{navlink} YAML
header element as well as support for limited OpenGraph header elements
(see the example Rmd) which will be used in header of the generated file.

}
When you use RStudio to create a new R Markdown document, select "From
Template" and choose one of these templates.
}
\author{
Ian Kloo (@iankloo)
}

